+++ ui/src/main/java/com/kotcrab/vis/ui/widget/file/internal/DirsSuggestionPopup.java
@@ -20,15 +20,23 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.files.FileHandle;
 import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.Stage;
+import com.badlogic.gdx.scenes.scene2d.actions.Actions;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.utils.Array;
 import com.kotcrab.vis.ui.widget.MenuItem;
 import com.kotcrab.vis.ui.widget.VisTextField;
 import com.kotcrab.vis.ui.widget.file.FileChooser;
 
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
+
 /** @author Kotcrab */
 public class DirsSuggestionPopup extends AbstractSuggestionPopup {
 	private final VisTextField pathField;
+
+	private ExecutorService listDirExecutor = Executors.newSingleThreadExecutor(new ServiceThreadFactory("FileChooserListDirThread"));
+	private Future<?> listDirFuture;
 
 	public DirsSuggestionPopup (FileChooser chooser, VisTextField pathField) {
 		super(chooser);
@@ -40,50 +48,73 @@ public class DirsSuggestionPopup extends AbstractSuggestionPopup {
 			remove();
 			return;
 		}
-
-		int suggestions = createDirSuggestions(width);
-		if (suggestions == 0) {
-			remove();
-			return;
-		}
-
-		showMenu(stage, pathField);
+		createDirSuggestions(stage, width);
 	}
 
-	private int createDirSuggestions (float width) {
-		clearChildren();
-		int suggestions = 0;
+	private void createDirSuggestions (final Stage stage, final float width) {
+		final String pathFieldText = pathField.getText();
+		//quiet period before listing files takes too long and popup will be removed
+		addAction(Actions.sequence(Actions.delay(0.2f, Actions.removeActor())));
 
-		FileHandle enteredDir = Gdx.files.absolute(pathField.getText());
-		String partialPath = "";
-		if (enteredDir.exists() == false) {
-			partialPath = enteredDir.name();
-			enteredDir = enteredDir.parent();
-		}
-
-		for (final FileHandle file : enteredDir.list(chooser.getFileFilter())) {
-			if (file.exists() == false || file.isDirectory() == false) continue;
-			if (file.name().startsWith(partialPath) == false || file.name().equals(partialPath)) continue;
-
-			MenuItem item = createMenuItem(file.path());
-			item.getLabel().setEllipsis(true);
-			item.getLabelCell().width(width);
-			addItem(item);
-
-			item.addListener(new ChangeListener() {
-				@Override
-				public void changed (ChangeEvent event, Actor actor) {
-					chooser.setDirectory(file, FileChooser.HistoryPolicy.ADD);
+		if (listDirFuture != null) listDirFuture.cancel(true);
+		listDirFuture = listDirExecutor.submit(new Runnable() {
+			@Override
+			public void run () {
+				FileHandle enteredDir = Gdx.files.absolute(pathFieldText);
+				final FileHandle listDir;
+				final String partialPath;
+				if (enteredDir.exists()) {
+					listDir = enteredDir;
+					partialPath = "";
+				} else {
+					listDir = enteredDir.parent();
+					partialPath = enteredDir.name();
 				}
-			});
 
-			suggestions++;
-			if (suggestions == MAX_SUGGESTIONS) {
-				break;
+				final FileHandle[] files = listDir.list(chooser.getFileFilter());
+				if (Thread.currentThread().isInterrupted()) return;
+				Gdx.app.postRunnable(new Runnable() {
+					@Override
+					public void run () {
+						clearChildren();
+						clearActions();
+						int suggestions = 0;
+
+						for (final FileHandle file : files) {
+							if (file.exists() == false || file.isDirectory() == false) continue;
+							if (file.name().startsWith(partialPath) == false || file.name().equals(partialPath))
+								continue;
+
+							MenuItem item = createMenuItem(file.path());
+							item.getLabel().setEllipsis(true);
+							item.getLabelCell().width(width - 20);
+							addItem(item);
+
+							item.addListener(new ChangeListener() {
+								@Override
+								public void changed (ChangeEvent event, Actor actor) {
+									chooser.setDirectory(file, FileChooser.HistoryPolicy.ADD);
+								}
+							});
+
+							suggestions++;
+							if (suggestions == MAX_SUGGESTIONS) {
+								break;
+							}
+						}
+
+						if (suggestions == 0) {
+							remove();
+							return;
+						}
+
+						showMenu(stage, pathField);
+						setWidth(width);
+						layout();
+					}
+				});
 			}
-		}
-
-		return suggestions;
+		});
 	}
 
 	public void showRecentDirectories (Stage stage, Array<FileHandle> recentDirectories, float width) {
@@ -92,8 +123,9 @@ public class DirsSuggestionPopup extends AbstractSuggestionPopup {
 			remove();
 			return;
 		}
-
 		showMenu(stage, pathField);
+		setWidth(width);
+		layout();
 	}
 
 	private int createRecentDirSuggestions (Array<FileHandle> files, float width) {
@@ -104,7 +136,7 @@ public class DirsSuggestionPopup extends AbstractSuggestionPopup {
 
 			MenuItem item = createMenuItem(file.path());
 			item.getLabel().setEllipsis(true);
-			item.getLabelCell().width(width);
+			item.getLabelCell().width(width - 20);
 			addItem(item);
 
 			item.addListener(new ChangeListener() {
@@ -123,3 +155,4 @@ public class DirsSuggestionPopup extends AbstractSuggestionPopup {
 		return suggestions;
 	}
 }
+